+++ packages/effect/src/optic/index.ts
@@ -11,8 +11,3 @@ export * as AST from "./AST.ts"
  * @since 4.0.0
  */
 export * as Optic from "./Optic.ts"
-
-/**
- * @since 4.0.0
- */
-export * as Optic2 from "./Optic2.ts"
