+++ packages/effect/src/schema/ToOptic.ts
@@ -2,8 +2,8 @@
  * @since 4.0.0
  */
 
-import * as AST from "../optic/AST.ts"
-import * as Optic from "../optic/Optic2.ts"
+import * as Op from "../optic/AST.ts"
+import * as Optic from "../optic/Optic.ts"
 import type * as Schema from "./Schema.ts"
 import * as Serializer from "./Serializer.ts"
 import * as ToParser from "./ToParser.ts"
@@ -15,7 +15,7 @@ export function makeIso<T, Iso>(codec: Schema.Optic<T, Iso>): Optic.Iso<T, Iso> 
   const serializer = Serializer.iso(codec)
   const get = ToParser.encodeSync(serializer)
   const set = ToParser.decodeSync(serializer)
-  return Optic.make(new AST.Iso(get, set))
+  return Optic.make(new Op.Iso(get, set))
 }
 
 /**
