+++ packages/effect/test/optic/Optic.test.ts
@@ -1,42 +1,51 @@
-import { Option } from "effect/data"
-import { Optic } from "effect/optic"
+import { AST, Optic } from "effect/optic"
 import { Check } from "effect/schema"
 import { describe, it } from "vitest"
-import {
-  assertFailure,
-  assertNone,
-  assertSome,
-  assertSuccess,
-  deepStrictEqual,
-  strictEqual,
-  throws
-} from "../utils/assert.ts"
+import { assertFailure, assertSuccess, assertTrue, deepStrictEqual, strictEqual } from "../utils/assert.ts"
+
+const addOne = (n: number) => n + 1
 
 describe("Optic", () => {
+  describe("AST", () => {
+    const iso = new AST.Iso<1, 2>(() => 2, () => 1)
+
+    it("composing an identity with another ast should return the other ast", () => {
+      const path = new AST.Path(["a"])
+      strictEqual(AST.compose(AST.identity, path), path)
+    })
+
+    it("composing two path asts should return a path ast with the two paths concatenated", () => {
+      const path1 = new AST.Path(["a"])
+      const path2 = new AST.Path(["b"])
+      const composed = AST.compose(AST.compose(iso, path1), path2)
+      assertTrue(composed._tag === "Composition")
+      strictEqual(composed.ops.length, 2)
+      strictEqual(composed.ops[0], iso)
+      deepStrictEqual(composed.ops[1], new AST.Path(["a", "b"]))
+    })
+
+    it("composing two checks asts should return a checks ast with the two checks concatenated", () => {
+      const checks1 = new AST.Checks([Check.positive()])
+      const checks2 = new AST.Checks([Check.int()])
+      const composed = AST.compose(AST.compose(iso, checks1), checks2)
+      assertTrue(composed._tag === "Composition")
+      strictEqual(composed.ops.length, 2)
+      strictEqual(composed.ops[0], iso)
+      deepStrictEqual(composed.ops[1], new AST.Checks([...checks1.checks, ...checks2.checks]))
+    })
+  })
+
   describe("Iso", () => {
     it("id", () => {
-      const iso = Optic.id<string>()
+      const iso = Optic.id<number>()
 
-      strictEqual(iso.get("a"), "a")
-      strictEqual(iso.set("b"), "b")
+      strictEqual(iso.get(1), 1)
+      strictEqual(iso.set(1), 1)
+      strictEqual(iso.modify(addOne)(1), 2)
     })
   })
 
   describe("Lens", () => {
-    it("get", () => {
-      type S = { readonly a: number }
-      const optic = Optic.id<S>().key("a")
-
-      strictEqual(optic.get({ a: 1 }), 1)
-    })
-
-    it("replace", () => {
-      type S = { readonly a: number }
-      const optic = Optic.id<S>().key("a")
-
-      deepStrictEqual(optic.replace(2, { a: 1 }), { a: 2 })
-    })
-
     describe("key", () => {
       describe("Struct", () => {
         it("required key", () => {
@@ -45,49 +54,108 @@ describe("Optic", () => {
 
           strictEqual(optic.get({ a: 1 }), 1)
           deepStrictEqual(optic.replace(2, { a: 1 }), { a: 2 })
+          deepStrictEqual(optic.modify(addOne)({ a: 1 }), { a: 2 })
         })
 
-        it("nested key", () => {
-          type S = { readonly a: { readonly b: number } }
+        it("nested required keys", () => {
+          type S = {
+            readonly a: {
+              readonly b: number
+            }
+          }
           const optic = Optic.id<S>().key("a").key("b")
 
           strictEqual(optic.get({ a: { b: 1 } }), 1)
           deepStrictEqual(optic.replace(2, { a: { b: 1 } }), { a: { b: 2 } })
+          deepStrictEqual(optic.modify(addOne)({ a: { b: 1 } }), { a: { b: 2 } })
         })
       })
 
-      it("Tuple", () => {
-        type S = [number]
-        const optic = Optic.id<S>().key(0)
+      describe("Tuple", () => {
+        it("required element", () => {
+          type S = readonly [number]
+          const optic = Optic.id<S>().key(0)
 
-        strictEqual(optic.get([1]), 1)
-        deepStrictEqual(optic.replace(2, [1]), [2])
+          strictEqual(optic.get([1]), 1)
+          deepStrictEqual(optic.replace(2, [1]), [2])
+          deepStrictEqual(optic.modify(addOne)([1]), [2])
+        })
+
+        it("nested required element", () => {
+          type S = readonly [string, readonly [number]]
+          const optic = Optic.id<S>().key(1).key(0)
+
+          strictEqual(optic.get(["a", [1]]), 1)
+          deepStrictEqual(optic.replace(2, ["a", [1]]), ["a", [2]])
+          deepStrictEqual(optic.modify(addOne)(["a", [1]]), ["a", [2]])
+        })
+      })
+
+      it("Mixed structs and tuples", () => {
+        type S = {
+          readonly a: readonly [number]
+        }
+        const optic = Optic.id<S>().key("a").key(0)
+
+        strictEqual(optic.get({ a: [1] }), 1)
+        deepStrictEqual(optic.replace(2, { a: [1] }), { a: [2] })
+        deepStrictEqual(optic.modify(addOne)({ a: [1] }), { a: [2] })
+      })
+    })
+  })
+
+  describe("Prism", () => {
+    describe("check", () => {
+      it("single check", () => {
+        type S = number
+        const optic = Optic.id<S>().check(Check.positive())
+        assertSuccess(optic.getResult(1), 1)
+        assertFailure(optic.getResult(0), `Expected a value greater than 0, got 0`)
+        strictEqual(optic.set(1), 1)
+        strictEqual(optic.set(0), 0)
+        deepStrictEqual(optic.modify(addOne)(1), 2)
+        deepStrictEqual(optic.modify(addOne)(0), 0)
+      })
+
+      it("multiple checks", () => {
+        type S = number
+        const optic = Optic.id<S>().check(Check.int(), Check.positive())
+        assertSuccess(optic.getResult(1), 1)
+        assertFailure(optic.getResult(0), `Expected a value greater than 0, got 0`)
+        assertFailure(optic.getResult(1.1), `Expected an integer, got 1.1`)
+        assertFailure(
+          optic.getResult(-1.1),
+          `Expected an integer, got -1.1
+Expected a value greater than 0, got -1.1`
+        )
+        deepStrictEqual(optic.modify(addOne)(1), 2)
+        deepStrictEqual(optic.modify(addOne)(0), 0)
+        deepStrictEqual(optic.modify(addOne)(1.1), 1.1)
+        deepStrictEqual(optic.modify(addOne)(-1.1), -1.1)
       })
     })
 
-    describe("optionalKey", () => {
-      it("Struct", () => {
-        type S = { readonly a?: number }
-        const optic = Optic.id<S>().optionalKey("a")
+    it("refine", () => {
+      type B = { readonly _tag: "b"; readonly b: number }
+      type S = { readonly _tag: "a"; readonly a: string } | B
+      const optic = Optic.id<S>().refine(
+        Check.makeRefineByGuard((s: S): s is B => s._tag === "b", { title: `"b" tag` })
+      ).key("b")
 
-        strictEqual(optic.get({ a: 1 }), 1)
-        strictEqual(optic.get({}), undefined)
-        deepStrictEqual(optic.replace(2, { a: 1 }), { a: 2 })
-        deepStrictEqual(optic.replace(2, {}), { a: 2 })
-        deepStrictEqual(optic.replace(undefined, { a: 1 }), {})
-      })
+      assertSuccess(optic.getResult({ _tag: "b", b: 1 }), 1)
+      assertFailure(optic.getResult({ _tag: "a", a: "value" }), `Expected "b" tag, got {"_tag":"a","a":"value"}`)
+      deepStrictEqual(optic.modify(addOne)({ _tag: "a", a: "value" }), { _tag: "a", a: "value" })
+      deepStrictEqual(optic.modify(addOne)({ _tag: "b", b: 1 }), { _tag: "b", b: 2 })
+    })
 
-      it("Tuple", () => {
-        type S = readonly [number?, number?]
-        const optic = Optic.id<S>().optionalKey(0)
+    it("tag", () => {
+      type S = { readonly _tag: "a"; readonly a: string } | { readonly _tag: "b"; readonly b: number }
+      const optic = Optic.id<S>().tag("b").key("b")
 
-        strictEqual(optic.get([1]), 1)
-        strictEqual(optic.get([]), undefined)
-        deepStrictEqual(optic.replace(2, [1]), [2])
-        deepStrictEqual(optic.replace(2, []), [2])
-        deepStrictEqual(optic.replace(undefined, [1]), [])
-        throws(() => optic.replace(undefined, [1, 2]), `Cannot remove element at index 0`)
-      })
+      assertSuccess(optic.getResult({ _tag: "b", b: 1 }), 1)
+      assertFailure(optic.getResult({ _tag: "a", a: "value" }), `Expected "b" tag, got "a"`)
+      deepStrictEqual(optic.modify(addOne)({ _tag: "a", a: "value" }), { _tag: "a", a: "value" })
+      deepStrictEqual(optic.modify(addOne)({ _tag: "b", b: 1 }), { _tag: "b", b: 2 })
     })
 
     describe("at", () => {
@@ -97,8 +165,8 @@ describe("Optic", () => {
 
         assertSuccess(optic.getResult({ a: 1, b: 2 }), 1)
         assertFailure(optic.getResult({ b: 2 }), `Key "a" not found`)
-        assertFailure(optic.replaceResult(2, { b: 2 }), `Key "a" not found`)
-        assertSuccess(optic.replaceResult(2, { a: 1, b: 2 }), { a: 2, b: 2 })
+        deepStrictEqual(optic.replace(2, { b: 2 }), { b: 2 })
+        deepStrictEqual(optic.replace(2, { a: 1, b: 2 }), { a: 2, b: 2 })
       })
 
       it("Array", () => {
@@ -107,107 +175,20 @@ describe("Optic", () => {
 
         assertSuccess(optic.getResult([1, 2]), 1)
         assertFailure(optic.getResult([]), `Key 0 not found`)
-        assertFailure(optic.replaceResult(2, []), `Key 0 not found`)
-        assertSuccess(optic.replaceResult(3, [1, 2]), [3, 2])
+        deepStrictEqual(optic.replace(2, []), [])
+        deepStrictEqual(optic.replace(3, [1, 2]), [3, 2])
       })
     })
   })
 
-  describe("Prism", () => {
-    it("getOption", () => {
-      const optic = Optic.id<number>().check(Check.positive())
-
-      assertSome(optic.getOption(2), 2)
-      assertNone(optic.getOption(0))
-    })
-
-    it("set", () => {
-      const optic = Optic.id<number>().check(Check.positive())
-
-      strictEqual(optic.set(2), 2)
-      strictEqual(optic.set(0), 0)
-    })
-
-    it("check", () => {
-      const optic = Optic.id<number>().check(Check.positive())
-
-      assertSuccess(optic.getResult(1), 1)
-      assertFailure(optic.getResult(0), `Expected a value greater than 0, got 0`)
-      assertSuccess(optic.replaceResult(2, 2), 2)
-      assertSuccess(optic.replaceResult(0, 0), 0)
-    })
-  })
-
   describe("Optional", () => {
-    it("getOption", () => {
-      type S = { readonly a: number }
-      const optic = Optic.id<S>().key("a").check(Check.positive())
-
-      assertSome(optic.getOption({ a: 1 }), 1)
-      assertNone(optic.getOption({ a: 0 }))
-    })
-
-    it("replace", () => {
-      type S = { readonly a: number }
-      const optic = Optic.id<S>().key("a").check(Check.positive())
-
-      deepStrictEqual(optic.replace(2, { a: 1 }), { a: 2 })
-      deepStrictEqual(optic.replace(0, { a: 1 }), { a: 0 })
-    })
-
     it("key & check", () => {
       type S = { readonly a: number }
       const optic = Optic.id<S>().key("a").check(Check.positive())
-
       assertSuccess(optic.getResult({ a: 1 }), 1)
       assertFailure(optic.getResult({ a: 0 }), `Expected a value greater than 0, got 0`)
-      assertSuccess(optic.replaceResult(2, { a: 1 }), { a: 2 })
-      assertSuccess(optic.replaceResult(0, { a: 1 }), { a: 0 })
+      deepStrictEqual(optic.replace(2, { a: 1 }), { a: 2 })
+      deepStrictEqual(optic.replace(2, { a: 0 }), { a: 0 })
     })
-  })
-
-  it("modify", () => {
-    type S = { readonly a: number }
-    const optic = Optic.id<S>().key("a").check(Check.positive())
-    const f = optic.modify((a) => a + 1)
-
-    deepStrictEqual(f({ a: 1 }), { a: 2 })
-    deepStrictEqual(f({ a: 0 }), { a: 0 }) // getResult fails
-  })
-
-  it("refine", () => {
-    const optic = Optic.id<Option.Option<number>>().refine(Check.some())
-
-    assertSuccess(optic.getResult(Option.some(1)), Option.some(1))
-    assertFailure(optic.getResult(Option.none()), `Expected a Some value, got none()`)
-  })
-
-  it("tag", () => {
-    type S = { readonly _tag: "a"; readonly a: string } | { readonly _tag: "b"; readonly b: number }
-    const optic = Optic.id<S>().tag("a").key("a")
-
-    assertSuccess(optic.getResult({ _tag: "a", a: "value" }), "value")
-    assertFailure(optic.getResult({ _tag: "b", b: 1 }), `Expected "a" tag, got "b"`)
-  })
-
-  it("some", () => {
-    const optic = Optic.id<Option.Option<number>>().compose(Optic.some())
-
-    assertSuccess(optic.getResult(Option.some(1)), 1)
-    assertFailure(optic.getResult(Option.none()), `Expected a Some value, got none()`)
-
-    assertSuccess(optic.replaceResult(2, Option.some(1)), Option.some(2))
-    assertSuccess(optic.replaceResult(2, Option.none()), Option.some(2))
-  })
-
-  it("charAt", () => {
-    const optic = Optic.id<string>().compose(Optic.charAt(0))
-
-    assertSuccess(optic.getResult("abc"), "a")
-    assertFailure(optic.getResult(""), `Missing character at index 0`)
-    assertSuccess(optic.replaceResult("d", "abc"), "dbc")
-    assertFailure(optic.replaceResult("", "abc"), `Expected a single character, got ""`)
-    assertFailure(optic.replaceResult("de", "abc"), `Expected a single character, got "de"`)
-    assertFailure(optic.replaceResult("d", ""), `Missing character at index 0`)
   })
 })
